==> 09_word_cloud.py <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from matplotlib import font_manager, rc
import collections
from wordcloud import WordCloud
from PIL import Image


# 폰트 지정
font_path = './malgun.ttf'
font_name = font_manager.FontProperties(fname=font_path).get_name()
plt.rc('font', family='NanumBarunGothic')


# 데이터 로드
df = pd.read_csv('./token_reviews_2017_2022.csv')


# 영화 제목으로 변수 지정
words = df[df['titles']=='1987 (1987:When the Day Comes)']['reviews']
# # 인덱스로 변수 지정
# words = df.iloc[2471][-1]

# print(words)            # 2471    영화 데스 숏컷 메인 예고편 공포 스릴러 판타지 평화롭다 오후 명의 학생 태운 버스...
# print(words.iloc[0])    # 영화 데스 숏컷 메인 예고편 공포 스릴러 판타지 평화롭다 오후 명의 학생 태운 버스...
words = words.iloc[0].split()
print(words)    # ['영화', '데스', '숏컷', '메인', '예고편', '공포', '스릴러', '판타지', '평화롭다', '오후', '명의', '학생', '태운', '버스', '운전', '세다', '도로', '장애물', '움직이다', '선택', '낯선', '갈림길', '방치', '사슴', '발견', '버스', '내리다', '세다', '갑작스레', '덮다', '치다', '옥수', '인질', '잡히다', '불청객', '탑승', '버스', '계속', '달린다', '두운', '터널', '멈추다', '버리다', '버스', '예고', '없이', '고립', '믿다', '생물체', '나타나다', '벗어나다', '공포', '터널', '갇히다', '출구', '영화', '데스', '숏컷', '사건', '진상', '영상', '공포', '스릴러', '판타지', '예고편', '감독', '알레', '시오', '오리', '출연', '케인', '안드레이', '클로드', '렌스', '앤더슨', '영화', '데드', '숏컷', '정보', '영상', '유익하다', '구독', '좋다', '누르다', '구독', '크다', '데스', '숏컷', '장르', '판타지', '공포', '스릴러', '줄거리', '스쿨버스', '명의', '남여', '태우다', '이동', '도로', '한가운데', '가로막', '장애물', '나타나다', '도저히', '치우다', '크기', '없이', '자주', '다니다', '갈림길', '들어서다', '평점', '탈주', '범의', '등장', '생물체', '등장', '특별하다', '미련', '버리다', '너무도', '허무하다', '엔딩', '잃다', '만들다', '영화', '결말', '데스', '숏컷', '평화롭다', '오후', '명의', '학생', '태운', '버스', '운전', '세다', '도로', '장애물', '움직이다']


# 워드클라우드 만들기
worddict = collections.Counter(words)
worddict = dict(worddict)
print(worddict) # {'영화': 4, '데스': 4, '숏컷': 5, '메인': 1, '예고편': 2, '공포': 4, '스릴러': 3, '판타지': 3, '평화롭다': 2, '오후': 2, '명의': 3, '학생': 2, '태운': 2, '버스': 5, '운전': 2, '세다': 3, '도로': 3, '장애물': 3, '움직이다': 2, '선택': 1, '낯선': 1, '갈림길': 2, '방치': 1, '사슴': 1, '발견': 1, '내리다': 1, '갑작스레': 1, '덮다': 1, '치다': 1, '옥수': 1, '인질': 1, '잡히다': 1, '불청객': 1, '탑승': 1, '계속': 1, '달린다': 1, '두운': 1, '터널': 2, '멈추다': 1, '버리다': 2, '예고': 1, '없이': 2, '고립': 1, '믿다': 1, '생물체': 2, '나타나다': 2, '벗어나다': 1, '갇히다': 1, '출구': 1, '사건': 1, '진상': 1, '영상': 2, '감독': 1, '알레': 1, '시오': 1, '오리': 1, '출연': 1, '케인': 1, '안드레이': 1, '클로드': 1, '렌스': 1, '앤더슨': 1, '데드': 1, '정보': 1, '유익하다': 1, '구독': 2, '좋다': 1, '누르다': 1, '크다': 1, '장르': 1, '줄거리': 1, '스쿨버스': 1, '남여': 1, '태우다': 1, '이동': 1, '한가운데': 1, '가로막': 1, '도저히': 1, '치우다': 1, '크기': 1, '자주': 1, '다니다': 1, '들어서다': 1, '평점': 1, '탈주': 1, '범의': 1, '등장': 2, '특별하다': 1, '미련': 1, '너무도': 1, '허무하다': 1, '엔딩': 1, '잃다': 1, '만들다': 1, '결말': 1}


wordcloud_img = WordCloud(background_color='white',
                          max_words=2000,
                          font_path=font_path).generate_from_frequencies(worddict)

plt.figure(figsize=(12,12))
plt.imshow(wordcloud_img, interpolation='bilinear')
plt.axis('off')
plt.show()
